requirements.txt
aioice==0.10.2
aiortc==1.14.0
altair==6.0.0
attrs==25.4.0
av==16.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dnspython==2.8.0
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-crc32c==1.8.0
idna==3.11
ifaddr==0.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
pyee==13.0.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
streamlit-webrtc==0.64.5
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
